feat: Add TensorBoard summaries
Added TensorBoard summaries to the loss, learning rate, all metrics and all decoder variables.

Also refactored the metrics and evaluation parts of the code to make it cleaner and easier to add new metrics.

For the training, you can now select what metrics the model shall be evaluated on and how frequently, and you can choose on which dataset (training or validation) all metrics should be evaluated.