fix: Fix metric value comparisons for  saving during training
Before the fix, all metric values were erroneously accepted to be better than their previous best value if they were lower. Of course this is only true for metrics that are to be minimized. This has now been corrected: For metrics that are to be maximized, such as mean_iou or accuracy, a new metric value is now accepted to be better than its best previous value if it is *larger*.